v0.93 (EZ Wild Apricot Web Designer) (#48)
* Change isInEditMode() to use class for detection

* Hide language toggle button if page loaded in widget mode

* Toggel second language if keyword present in URL

* Initial commit

* Rename wildapricot-textmanager-config-sample.csv to wildapricot-textmanager-config.csv

renamed the config file without the word sample, so it works right away without renaming

* added standard French translation configuration file

this is a standard French configuration file that contains over 600 standard strings for WildApricot. It should work with any theme, but needs some testing in the wild. It is installed on newpathconsulting.wildapricot.org as a testbed.

* updated with place holder v0.93 version for notes

- added standard French translation file notes
- renamed WATM to EZ Wild Apricot Web Designer

* updated 2 links

- added link for official Google docs on  How to use Chrome Developer Tools
- removed video preview on Chrome Developer Tools to find CSS classes or IDs

* added support for loading widgets in 3rd party CMS

- added docs for integrating with 3rd party cms systems
- updated link for dev tools and removed thumbnail from CSS video

* added reminder to rename french translation file

added reminder to save as wildapricot-textmanager-config.csv when using the French translation file

* updated website link for ez web designer

updated link in opening paragraph to reflect the new url for EZ Web Designer

* added comment on path in JS install code snippet

note to make sure the path in the install code snippet uses a defined path that can be changed if the files are uploaded into a different path

* updated for v0.93

renamed EZ Web Designer for WildApricot

Co-authored-by: Matthieu Disna <info@mdisna.dev>
Co-authored-by: mDisna <mdisna@gmail.com>